Resolve "Merge MATLAB Toolbox Windows and Linux"

Closes #155

See merge request common/sw/urx!147

+++ urx/utils/io/reader_impl.h
@@ -308,7 +308,7 @@ class ReaderBase {
 
       field = urx::utils::rawDataFactory(urx::utils::io::enums::h5PredTypeToDataType(datatype),
                                          dimension[1] == 1 ? SamplingType::RF : SamplingType::IQ,
-                                         dimension[0]);
+                                         static_cast<size_t>(dimension[0]));
 
       dataset.read(field->getBuffer(), datatype);
     } else {
